+++ tests/models/test_property_reference.py
@@ -487,6 +487,36 @@ class TestMultiReferenceProperty(TestBetamax):
 
         self.assertEqual(ref_to_wheel.scope_id, self.project.id)
 
+    def test_property_clear_selected_part_model(self):
+        # setUp
+        wheel_model = self.project.model('Wheel')
+        self.ref_prop_model.value = [wheel_model]
+
+        self.assertTrue(wheel_model.id == self.ref_prop_model.value[0].id)
+
+        self.ref_prop_model.value = None
+
+        # testing
+        self.assertIsNone(self.ref_prop_model.value)
+
+    def test_property_clear_referenced_part_instances(self):
+        # setUp
+        wheel_model = self.project.model('Wheel')
+        self.ref_prop_model.value = [wheel_model]
+
+        wheel_instances = wheel_model.instances()
+        wheel_instances_list = [instance.id for instance in wheel_instances]
+
+        # set ref value
+        self.ref.value = wheel_instances_list
+
+        # clear referenced part instances
+        self.ref.value = None
+
+        # testing
+        self.assertIsNone(self.ref.value)
+        self.assertIsNotNone(self.ref_prop_model.value)
+
     def test_copy_reference_property_with_options(self):
         # setUp
         copied_ref_property = self.ref_prop_model.copy(target_part=self.part_model,